Bug 1821799 - Make `AutoTrackDOMPoint` clear tracking point if the result gets lost from the document r=m_kato
It should not meet such situation if only editor updates the DOM tree.
Therefore, the result should be cleared in that case and all users should return
`NS_ERROR_EDITOR_UNEXPECTED_DOM_TREE` in this case.

Depends on D180785

Differential Revision: https://phabricator.services.mozilla.com/D180786

+++ editor/libeditor/HTMLEditorDeleteHandler.cpp
@@ -3498,6 +3498,11 @@ Result<EditActionResult, nsresult> HTMLEditor::AutoDeleteRangesHandler::
           "CaretPoint::SuggestCaretPointTo() failed, but ignored");
     }
   }
+
+  if (NS_WARN_IF(!aRangesToDelete.FirstRangeRef()->IsPositioned())) {
+    return Err(NS_ERROR_EDITOR_UNEXPECTED_DOM_TREE);
+  }
+
   nsresult rv =
       mDeleteRangesHandler->DeleteUnnecessaryNodesAndCollapseSelection(
           aHTMLEditor, aDirectionAndAmount,
